Explicitly Disable Style/Encoding (support RuboCop >= 0.50.0)
As of RuboCop v0.50.0, the Style/Encoding cop is now enabled by default
and configured to remove the encoding magic-comment. This deviates from
the expected behavior in specs and changes the behavior of this gem.

Disabling this cop will ensure that fix_db_schema_conflicts will output
consistently regardless of RuboCop version.

Change in RuboCop:
  https://github.com/bbatsov/rubocop/pull/4445